Restructure FA backend to be a bit cleaner, use structlog for logging
