+++ flowauth/backend/flowauth/admin.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 from cryptography.hazmat.primitives import serialization
-from flask import Blueprint, jsonify
+from flask import Blueprint, jsonify, request
 
 from flask_login import login_required
 from flask_principal import Permission, RoleNeed
@@ -53,5 +53,5 @@ def get_public_key():
         .decode()
         .strip()
     )
-    current_app.logger.debug(key)
+    current_app.logger.debug("Got public key", key=key)
     return jsonify({"public_key": key})
