+++ flowauth/backend/tests/test_access_reflects_server_limits.py
@@ -1,13 +1,6 @@
-from flowauth import (
-    Group,
-    GroupServerPermission,
-    GroupServerTokenLimits,
-    Server,
-    ServerCapability,
-    User,
-    datetime,
-    db,
-)
+import datetime
+
+from flowauth.models import Group, GroupServerPermission, GroupServerTokenLimits, Server, ServerCapability, User, db
 
 
 def test_disallow_right_on_server_disallows_for_group(app):
