+++ flowauth/backend/tests/test_commands.py
@@ -2,8 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from flowauth import add_admin_command
-from flowauth.models import User, db, init_db_command, demodata, Group
+from flowauth.cli import add_admin_command, demo_data, init_db_command
+from flowauth.models import Group, User, db
 
 
 def test_add_admin(app):
@@ -34,6 +34,7 @@ def test_add_admin_promotes(app):
         result = runner.invoke(
             add_admin_command, ["DUMMY_ADMINISTRATOR", "DUMMY_ADMINISTRATOR_PASSWORD"]
         )
+        assert result.exit_code == 0
         user = User.query.filter(User.username == "DUMMY_ADMINISTRATOR").first()
         assert user.is_admin
         assert user.password != original_password_hash
@@ -69,18 +70,19 @@ def test_init_db_force(app):
         assert len(User.query.all()) == 0
 
 
-def test_demo_data_only_sets_up_once(app, caplog):
+def test_demo_data_only_sets_up_once(app):
     """
     Demo data should only be created once.
     """
     with app.app_context():
         runner = app.test_cli_runner()
         app.config["DB_IS_SET_UP"].clear()
-        result = runner.invoke(demodata)
-        result = runner.invoke(demodata)
+        result = runner.invoke(demo_data)
+        result = runner.invoke(demo_data)
         assert len(User.query.all()) == 2
         assert len(Group.query.all()) == 3
-    assert "Database already set up by another worker, skipping." in caplog.text
+    assert False
+    # assert "Database already set up by another worker, skipping." in caplog.text
 
 
 def test_db_init_only_runs_once(app, caplog):
@@ -102,6 +104,6 @@ def test_demo_data(app):
     with app.app_context():
         runner = app.test_cli_runner()
         app.config["DB_IS_SET_UP"].clear()
-        result = runner.invoke(demodata)
+        result = runner.invoke(demo_data)
         assert len(User.query.all()) == 2
         assert len(Group.query.all()) == 3
